Attributes support evolution. Now layers support attributes too.
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@6419 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/tests/kits/net/new_stack/stack/dump.c
@@ -10,7 +10,7 @@ void dump_memory
 	(
 	const char *	prefix,
 	const void *	data,
-	uint32			len
+	size_t			len
 	)
 {
 	uint32	i,j;
