+++ src/tests/kits/net/new_stack/stack/dump.h
@@ -1,6 +1,8 @@
 #ifndef DUMP_H
 #define DUMP_H
 
-extern void dump_memory(const char *prefix, const void *data, uint32 len);
+#include <SupportDefs.h>
+
+extern void dump_memory(const char *prefix, const void *data, size_t len);
 
 #endif
